fixes in RPM building scripts

+++ twindb-backup.sh
@@ -1,11 +1,58 @@
 #!/usr/bin/env bash
 set -eu
 
+# Default settings
+# Source
+backup_dirs="/etc /root /home"
+backup_mysql=TRUE
+
+# Destination
+
+# Uncomment one
+backup_destination="s3"
+#backup_destination="ssh"
+
+# S3 destination settings
+AWS_ACCESS_KEY_ID="XXXXX"
+AWS_SECRET_ACCESS_KEY="YYYYY"
+AWS_DEFAULT_REGION="us-east-1"
+BUCKET="twindb-backups"
+
+# SSH destination settings
+backup_host=127.0.0.1
+ssh_user="ec2-user"
+backup_dir="/path/to/twindb-server-backups"
+
+# MySQL
+mysql_defaults_file=/etc/twindb/my.cnf
+
+# Retention
+hourly_copies=24
+daily_copies=7
+weekly_copies=4
+monthly_copies=12
+yearly_copies=3
+
+# Run intervals
+run_hourly="YES"
+run_daily="YES"
+run_weekly="YES"
+run_monthly="YES"
+run_yearly="YES"
+
+# EOF Default settings
+
 CONFIG=/etc/twindb/twindb-backup.cfg
 if test -f "${CONFIG}"
 then
     source ${CONFIG}
 fi
+
+export AWS_ACCESS_KEY_ID
+export AWS_SECRET_ACCESS_KEY
+export AWS_DEFAULT_REGION
+export BUCKET
+
 
 # Get time when script started
 start_time=`date +%s`
@@ -26,7 +73,12 @@ function get_ssh_key_path() {
 
 SSH="ssh $ssh_opt -l ${ssh_user} -i `get_ssh_key_path`"
 MYSQLDUMP="mysqldump"
-MYSQLDUMP_ARGS="--defaults-file=${mysql_defaults_file} -A"
+if test -f "${mysql_defaults_file}"
+then
+    MYSQLDUMP_ARGS="--defaults-file='${mysql_defaults_file}' -A"
+else
+    MYSQLDUMP_ARGS=" -A"
+fi
 
 trap "rm -f ${pidfile}" INT
 trap "rm -f ${pidfile}" ERR
@@ -57,13 +109,22 @@ function check_ssh() {
     done
 }
 
+function check_aws() {
+    aws s3 ls "s3://${BUCKET}" || aws s3 mb "s3://${BUCKET}"
+}
+
 function backup_mysql() {
     if [ "${backup_mysql}" = TRUE ]
     then
         mysql_backup_file=$1
         mysql_backup_dir="`dirname ${mysql_backup_file}`"
-        ${SSH} ${backup_host} mkdir -p "${mysql_backup_dir}"
-        ${MYSQLDUMP} ${MYSQLDUMP_ARGS} | gzip -c - | ${SSH} ${backup_host} "cat - > $mysql_backup_file"
+        if [ ${backup_destination} == "ssh" ]
+        then
+            ${SSH} ${backup_host} mkdir -p "${mysql_backup_dir}"
+            ${MYSQLDUMP} ${MYSQLDUMP_ARGS} | gzip -c - | ${SSH} ${backup_host} "cat - > $mysql_backup_file"
+        else
+            ${MYSQLDUMP} ${MYSQLDUMP_ARGS} | gzip -c - | aws s3 cp - "s3://${BUCKET}/${mysql_backup_file}"
+        fi
     fi
 }
 
@@ -73,7 +134,6 @@ function backup_files() {
         desc=`echo ${d} | sed 's/\///'`
         desc=`echo ${desc} | sed 's/\//_/g'`
         f="$desc-`date +%F_%H-%M-%S`.tar.gz"
-        ${SSH} ${backup_host} mkdir -p "$1"
         case ${OSTYPE} in
             "darwin14")
                 TAR_ARGS="zcLf -"
@@ -82,7 +142,17 @@ function backup_files() {
                 TAR_ARGS="zchf -"
         esac
 
-        tar ${TAR_ARGS} "$d" 2>/dev/null | ${SSH} ${backup_host} "cat - > $1/$f"
+        if [ ${backup_destination} == "ssh" ]
+        then
+            ${SSH} ${backup_host} mkdir -p "$1"
+        fi
+
+        if [ ${backup_destination} == "ssh" ]
+        then
+            tar ${TAR_ARGS} "$d" 2>/dev/null | ${SSH} ${backup_host} "cat - > $1/$f"
+        else
+            tar ${TAR_ARGS} "$d" 2>/dev/null | aws s3 cp -  "s3://${BUCKET}/$1/$f"
+        fi
     done
 }
 
@@ -108,7 +178,10 @@ function cleanup_old_backups() {
             vlog "Warning: Unknown backup period $suffix"
             exit -1
     esac
-    ${SSH} ${backup_host} "find $1 -mmin +$mtime -type f -delete"
+    if [ "${backup_destination}" == "ssh" ]
+    then
+        ${SSH} ${backup_host} "find $1 -mmin +$mtime -type f -delete"
+    fi
 }
 
 if  (! test -z "$1") && (
@@ -124,9 +197,18 @@ else
     echo "    `basename $0` <hourly | daily | weekly | monthly | yearly>"
     exit -1
 fi
-mysql_backup_dir="${backup_dir}/`hostname`/${suffix}/mysql"
+
+if [ "${backup_destination}" == "ssh" ]
+then
+    mysql_backup_dir="${backup_dir}/${mysql_backup_dir}"
+    files_backup_dir="${backup_dir}/`hostname`/$suffix"
+else
+    mysql_backup_dir="${suffix}/`hostname`/mysql"
+    files_backup_dir="${suffix}/`hostname`"
+fi
 mysql_backup_file="${mysql_backup_dir}/mysql-`date +%F_%H-%M-%S`.sql.gz"
-files_backup_dir="${backup_dir}/`hostname`/$suffix"
+
+
 
 # Do not run backups if run_* is set and it is NO
 if [ "$suffix" == "hourly" ] && ! [ -z ${run_hourly+x} ] && [ "${run_hourly}" == "NO" ]
@@ -164,11 +246,16 @@ echo ${pid} 1>&8
 # wait random time between 0 and 600 seconds to avoid storm starts
 max_delay=600
 let wait_time=$RANDOM*$max_delay/32767
-sleep ${wait_time}
-check_ssh
-cleanup_old_backups "${files_backup_dir}"
-#try create
-${SSH} ${backup_host} true || (vlog "Can not ssh to ${backup_host}"; exit -1)
-${SSH} ${backup_host} mkdir -p "${backup_dir}/`hostname`"
+if [ ${backup_destination} == "ssh" ]
+then
+    sleep ${wait_time}
+    check_ssh
+    cleanup_old_backups "${files_backup_dir}"
+    #try create
+    ${SSH} ${backup_host} true || (vlog "Can not ssh to ${backup_host}"; exit -1)
+    ${SSH} ${backup_host} mkdir -p "${backup_dir}/`hostname`"
+else
+    check_aws
+fi
 backup_mysql "${mysql_backup_file}"
 backup_files "${files_backup_dir}"
